Improve installation instructions
The previous installation instructions showed the library being installed to the Arduino IDE installation folder. That is a very bad idea because anything installed to that location will be lost whenever you update to a new version of the Arduino IDE. The correct place to install a library is to the sketchbook folder. The Arduino IDE has a feature that makes library installation very easy and will automatically install the library to the correct location. Although this feature is called "Add .ZIP Library", it can also be used to install libraries from a folder rather than a .zip file.

I have also improved the formatting and fixed typos in the installation instructions.